requirements.txt
Flask==2.2.2
PyMuPDF==1.21.0
Pillow==9.1.0
pytesseract==0.3.9
